Delete old code
