--- Clouds/Sections/Week/Components/DayPicker/Models/DayPickerItem.swift
@@ -1,15 +1,0 @@
-//
-//  DayPickerItem.swift
-//  Clouds
-//
-//  Created by Lukas Romsicki on 2020-04-08.
-//  Copyright © 2020 Lukas Romsicki. All rights reserved.
-//
-
-import Foundation
-
-struct DayPickerItem: Equatable {
-    let date: Date
-    let hasDayConditions: Bool
-    let hasNightConditions: Bool
-}
